refactor(sse): split chatCore.ts pure helpers into chatCore/ modules (−561 LOC)
Extracts the pure, side-effect-free helpers out of the chatCore.ts god-file
into focused open-sse/handlers/chatCore/ modules (headers, logTruncation,
memoryExtraction, nonStreamingSse, passthroughToolNames, upstreamTimeouts),
each with its own unit test. Behavior-preserving — chatCore.ts re-imports them.

Re-synced cleanly onto release/v3.8.29: the prior branch tip carried a
stale-merge regression that re-bloated open-sse/config/freeModelCatalog.data.ts
(462 -> 4530 lines) and reverted the free-tier refresh + provider/docs changes.
This commit keeps ONLY the chatCore split (zero contamination).
